simple-chatbot/pyproject.toml
[project]
name = "simple-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.61",
    "langchain-ollama>=0.3.3",
    "langchain-tavily>=0.1.6",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.7",
    "langsmith>=0.3.42",
    "python-dotenv>=1.1.0",
]
